+++ doc/source/changelog.rst
@@ -17,7 +17,12 @@ v1.6.1
   * ``cgf["bokeh"]["update_event"] = False``: Bokeh won't update the plot with
     new data as dragging or zooming operations are performed.
 
-* Added label capability to ``plot_implicit`` with ``MatplotlibBackend``.
+* Improvements to ``MatplotlibBackend``:
+
+  * Added label capability to ``plot_implicit``.
+
+  * ``show()`` method now accepts keyword arguments. This is useful to detach
+    the plot from a non-interactive console. 
 
 * Added ``dots`` keyword argument to ``plot_piecewise`` to choose wheter to
   show circular markers on endpoints.